Refactor - split rasp.js and move it to rasp directory
